backend/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
Flask-CORS==5.0.0
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.10
h3==4.1.2
requests==2.32.3
python-telegram-bot==21.7

# Optional: Coqui TTS for radio comms (pip install coqui-tts torch)
# Set ENABLE_RADIO_TTS=true to use. Without it, frontend falls back to browser TTS.
